% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq.R
\name{fastq_from_template}
\alias{fastq_from_template}
\title{Create a fastq object from an excel template}
\usage{
fastq_from_template(fastq_template)
}
\description{
Create a fastq object from an excel template
}
